css-writing-modes-3/orthogonal-parent-width-min-content-002.html

<!DOCTYPE html>
<title>CSS Writing Modes Test: Parent's min-content width when orthogonal inline in inline-block</title>
<link rel="help" href="http://www.w3.org/TR/css-writing-modes-3/#orthogonal-flows">
<meta name="assert" content="Parent's min-content width in orthorgonal flows">
<meta name="flags" content="ahem">
<link rel="author" title="Koji Ishii" href="mailto:kojiishi@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.test {
    font:25px Ahem;
    margin-top:.2em;
}
.target {
    color:blue;
    writing-mode:vertical-rl;
}
.border {
    border-right:blue solid 1em;
}
.next {
    color:orange;
}
.inline-block {
    display:inline-block;
}
table {
    border-spacing:0px;
}
td {
    padding:0px;
}
</style>
<div id="container">
<p>Test passes if the <b>left</b> edge of the orange box <b>touches</b> the <b>right</b> edge of the blue box.
    There should be no spaces nor overlaps.
<p>If script is enabled, there should be one or more PASS and no FAIL.
<div class="test" title="Orthogonal inline in inline-block">
    <div class="inline-block"><span class="target">XX</span></div><span class="next">YY</span>
</div>
</div>
<script>
function runTests() {
    Array.prototype.forEach.call(document.querySelectorAll(".test"), function (node) {
        test(function () {
            var targetNode = node.querySelector(".target");
            var targetBounds = targetNode.getBoundingClientRect();
            var border = parseFloat(getComputedStyle(targetNode).borderRightWidth);
            assert_equals(targetBounds.height, (targetBounds.width - border) * 2, "writing-mode is vertical")
            var nextNode = node.querySelector(".next");
            var nextBounds = nextNode.getBoundingClientRect();
            assert_equals(nextBounds.left - targetBounds.right, 0, "the left edge of the orange box touches the right edge of the blue box");
        }, node.title);
    });
    if (window.testRunner)
        container.style.display = "none";
    done();
}

if (window.location.search == "?wait") {
    console.log("Sleeping 5 secs for debug");
    setup({explicit_done:true});
    window.setTimeout(runTests, 5000);
} else {
    runTests();
}
</script>
